requirements_20231204.txt
amqp==5.1.1
annotated-types==0.5.0
anyio==3.7.1
appnope==0.1.3
astor==0.8.1
asttokens==2.2.1
backcall==0.2.0
billiard==3.6.4.0
bitarray==2.6.0
bloom-filter==1.3.3
Brotli==1.1.0
celery==5.2.7
certifi==2023.7.22
charset-normalizer==2.1.1
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
contourpy==1.0.6
cvxopt==1.3.0
cycler==0.11.0
Cython==3.0.5
datasketch==1.6.4
decorator==5.1.1
deprecation==2.1.0
dill==0.3.6
dnspython==2.2.1
et-xmlfile==1.1.0
exceptiongroup==1.1.3
executing==1.2.0
fastapi==0.103.1
fonttools==4.38.0
formulaic==0.6.6
future==0.18.2
geocoder==1.38.1
gmpy2==2.1.2
graphviz==0.20.1
greenlet==3.0.1
idna==3.4
impyla==0.18.0
interface-meta==1.3.0
intervaltree==3.1.0
ipython==8.7.0
jedi==0.18.2
Jinja2==3.1.2
joblib==1.2.0
jsonpickle==3.0.1
kiwisolver==1.4.4
kombu==5.2.4
linearmodels==5.3
lxml==4.9.2
MarkupSafe==2.1.1
matplotlib==3.6.2
matplotlib-inline==0.1.6
mpmath==1.3.0
multiprocess==0.70.14
mypy-extensions==1.0.0
Naked==0.1.32
networkx==2.8.8
nltk==3.7
numpy==1.26.0
openpyxl==3.1.2
packaging==21.3
pandas==1.5.1
parso==0.8.3
pathos==0.3.0
patsy==0.5.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.3.0
pip==23.3.1
pm4py==2.3.3
pox==0.3.2
ppft==1.7.6.6
prompt-toolkit==3.0.32
ptyprocess==0.7.0
pure-eval==0.2.2
pure-sasl==0.6.2
pyahocorasick==2.0.0
pybind11==2.11.1
pycryptodome==3.16.0
pydantic==2.3.0
pydantic_core==2.6.3
pydotplus==2.0.2
Pygments==2.13.0
pyhdfe==0.2.0
pymongo==4.3.2
pyparsing==3.0.9
PySocks==1.7.1
python-dateutil==2.8.2
PyTrie==0.4.0
pytz==2022.6
pyvis==0.3.1
ratelim==0.1.6
regex==2022.10.31
requests==2.28.1
scikit-learn==1.1.3
scipy==1.11.3
setuptools==68.2.2
setuptools-scm==7.1.0
shellescape==3.8.1
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
SQLAlchemy==1.4.39
stack-data==0.6.2
starlette==0.27.0
statsmodels==0.14.0
StringDist==1.0.9
sympy==1.12
textdistance==4.5.0
threadpoolctl==3.1.0
thrift==0.16.0
thrift-sasl==0.4.3
tomli==2.0.1
torch==2.0.0.dev20230124
torchaudio==2.0.0.dev20230124
torchvision==0.15.0.dev20230125
tqdm==4.64.1
traitlets==5.7.1
typing_extensions==4.8.0
urllib3==1.26.18
vine==5.0.0
wcwidth==0.2.5
wheel==0.41.3
wrapt==1.16.0
xlrd==2.0.1
zhon==1.1.5
